feat: add upload_asset.sh script for uploading assets to GitHub releases
The `upload_asset.sh` script is added to the project. This script is responsible for uploading assets to GitHub releases. It requires the following environment variables to be set: `GITHUB_TOKEN`, `ASSET_NAME_PREFIX`, and `ASSET_DIR`. The script takes a `VERSION` parameter to specify the version of the asset.

The script retrieves the `UPLOAD_URL` using the `get_draft_release.sh` script and then iterates over an array of file extensions to upload each asset. It checks if the file exists and is not empty before uploading it. The script uses `curl` to make the API request to upload the asset.

fix: remove unnecessary download of upload_asset.sh in attach-artifact-release.yml workflow

The `attach-artifact-release.yml` workflow file had a redundant download of the `upload_asset.sh` script. This line has been removed to avoid duplication and unnecessary downloads.