chore(pom.xml): add proguard rules to keep specific classes from being obfuscated
The changes in this commit add ProGuard rules to the pom.xml file. These rules specify that certain classes should be kept from being obfuscated during the build process. The classes that are being kept include those from the following packages: io.netty, org.reactivestreams, com.google, com.fasterxml.jackson, org.apache, org.bouncycastle, and org.slf4j.

The reason for these changes is to ensure that the specified classes are not obfuscated, as they may be required for the proper functioning of the application or for compatibility with other libraries or frameworks. By keeping these classes, we can avoid potential issues that may arise from obfuscating them.